refactor: move AuctionDetailScreen to screens folder
