+++ mobile-app/app/screens/AppNavigator/screens/Auctions/AuctionNavigator.tsx
@@ -5,7 +5,7 @@ import { translate } from '@translations'
 import BigNumber from 'bignumber.js'
 import { NetworkDetails } from '../Settings/screens/NetworkDetails'
 import { AuctionsScreen } from './AuctionScreen'
-import { AuctionDetailScreen } from './AuctionDetailScreen'
+import { AuctionDetailScreen } from './screens/AuctionDetailScreen'
 import { LoanVaultLiquidated, LoanVaultLiquidationBatch } from '@defichain/whale-api-client/dist/api/loan'
 import { PlaceBidScreen } from './screens/PlaceBidScreen'
 import { ConfirmPlaceBidScreen } from './screens/ConfirmPlaceBidScreen'
